+++ src/main/java/online/kheops/auth_server/inbox/InboxInfoResponse.java
@@ -4,7 +4,6 @@ package online.kheops.auth_server.inbox;
 import org.jooq.Record;
 
 import javax.xml.bind.annotation.XmlElement;
-import java.math.BigDecimal;
 
 public class InboxInfoResponse {
 
@@ -23,7 +22,7 @@ public class InboxInfoResponse {
         this.numberOfStudies = (Integer) r.getValue("number_of_studies");
         this.numberOfSeries = (Integer) r.getValue("number_of_series");
         try {
-            this.numberOfInstances = ((BigDecimal) r.getValue("number_of_instances")).intValue();
+            this.numberOfInstances =  (Integer) r.getValue("number_of_instances");
         } catch(NullPointerException e) {
             this.numberOfInstances = 0;
         }
