+++ src/main/java/online/kheops/auth_server/study/Studies.java
@@ -22,7 +22,6 @@ import org.jooq.impl.DSL;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
 import javax.ws.rs.BadRequestException;
-import java.math.BigDecimal;
 import java.sql.Connection;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -220,7 +219,7 @@ public class Studies {
             safeAttributeSetString(attributes, Tag.PatientSex, VR.CS, r.getValue(STUDIES.PATIENT_SEX.getName()).toString());
             safeAttributeSetString(attributes, Tag.StudyID, VR.SH, r.getValue(STUDIES.STUDY_ID.getName()).toString());
             attributes.setInt(Tag.NumberOfStudyRelatedSeries, VR.IS, ((Integer) r.getValue("count:" + SERIES.PK.getName())));
-            attributes.setInt(Tag.NumberOfStudyRelatedInstances, VR.IS, ((BigDecimal) r.getValue("sum:" + INSTANCES.PK.getName())).intValue());
+            attributes.setInt(Tag.NumberOfStudyRelatedInstances, VR.IS, (Integer) r.getValue("sum:" + INSTANCES.PK.getName()));
 
             //Tag Type (3) Optional
             if(qidoParams.includeStudyDescriptionField() && r.getValue(STUDIES.STUDY_DESCRIPTION.getName()) != null) {
